requirements.txt
streamlit
requests
openai
PyGithub